+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
-- [#79] Fix a bug where an installation failed if old pvcs stuck with terminating status. 
+- [#79] Fix a bug where an installation failed if old PVCs stuck with terminating status. 
 
 ## [v0.25.0] - 2023-02-17
 ### Added